[Refactor] useStringWidget composable
